Added email verification. Adding new tests

+++ internal/register/user.go
@@ -5,7 +5,7 @@ import (
 )
 
 type User struct {
-	UserId              int       `json:user_id`
+	UserId              int64     `json:user_id`
 	Username            string    `json:username`
 	NormalizedUsername  string    `json:normalized_username`
 	PasswordHash        string    `json:password_hash`
